MAR-3708. Rename helper function to getRandomTitle, update snapshots
