+++ client/components/core/modals/ModalEbook.vue
@@ -42,7 +42,7 @@
 import Modal from '@/components/core/Modal'
 import ReadForm from '@/components/Ebook/ReadForm'
 import SuccessMessage from '@/components/core/modals/SuccessMessage'
-import makeRandomTitle from '@/helpers/makeRandomTitle'
+import getRandomTitle from '@/helpers/getRandomTitle'
 
 export default {
   name: 'ModalEbook',
@@ -98,7 +98,7 @@ export default {
   },
 
   mounted() {
-    this.title = makeRandomTitle(this.sendPulseTemplateId)
+    this.title = getRandomTitle(this.sendPulseTemplateId)
   },
 
   methods: {
